success to add friends to table

+++ Models/Friend.cs
@@ -8,7 +8,7 @@ namespace Models
 {
     public class Friend
     {
-        public int FriendID { get; set; } // Changed from Bet to int
+        public long FriendID { get; set; } // long
         public long UserID_1 { get; set; }
         public long UserID_2 { get; set; }
         public string Status { get; set; } // "pending", "accepted", "rejected"
